pyproject.toml
[project]
name = "personal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai==1.12.1",
    "groq>=1.1.2",
    "openai>=2.30.0",
    "python-dotenv==1.1.0",
    "requests>=2.33.0",
]
